// src/app/models/ListaCuotas.ts
export class ListaCuotas {
    checked:boolean;
    IdPreDebito: number;
    IdCertificado: number;
    DebitarCuota: string;
    FechaPrevista: string;
    FechaCobro: string;
    IdEstado: number;
    CodigoEstado: string;
    Estado: string;
    PVPCobrado: number;
    IVANombre: string;
    Porcentaje: number;
  }